functions/requirements.txt
firebase-admin
firebase-functions
google-cloud-aiplatform
pyyaml
